pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "This application is designed to automate the creation and uploading of youtube videos to my channel and monetize my channel"
authors = [
    {name = "Shubs2002",email = "s.d.nakashe2002@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "google-auth (>=2.40.3,<3.0.0)",
    "google-auth-oauthlib (>=1.2.2,<2.0.0)",
    "google-auth-httplib2 (>=0.2.0,<0.3.0)",
    "google-api-python-client (>=2.178.0,<3.0.0)",
    "google-cloud-aiplatform (>=1.70.0,<2.0.0)",
    "google-genai (>=1.52.0,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "openai (>=1.107.3,<2.0.0)",
    "google (>=3.0.0,<4.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "moviepy (>=2.2.1,<3.0.0)",
    "imageio-ffmpeg (>=0.6.0,<0.7.0)",
    "pymongo (>=4.15.2,<5.0.0)",
    "opencv-python (>=4.12.0.88,<5.0.0.0)",
    "cloudinary (>=1.41.0,<2.0.0)",
    "rembg (>=2.0.59,<3.0.0)",
    "cryptography (>=43.0.0,<44.0.0)",
    "python-jose[cryptography] (>=3.3.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "email-validator (>=2.0.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dev = "src.dev:main"
prod = "src.main:main"